+++ wifi/services/wifi_standard/wifi_framework/wifi_toolkit/config/wifi_config_file_impl.h
@@ -338,8 +338,8 @@ int WifiConfigFileImpl<T>::SaveConfig()
         content = mEncry.encryptedPassword;
     }
 #endif
-    int ret = fwrite(content.c_str(), 1, content.length(), fp);
-    if (ret != (int)content.length()) {
+    size_t ret = fwrite(content.c_str(), 1, content.length(), fp);
+    if (ret != content.length()) {
         LOGE("Save config file: %{public}s, fwrite() failed!", mFileName.c_str());
     }
     (void)fflush(fp);
